login page 2/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
